+++ app/src/main/java/org/jellyfin/androidtv/auth/model/AccountManagerAccount.kt
@@ -6,5 +6,5 @@ data class AccountManagerAccount(
 	val id: UUID,
 	val server: UUID,
 	val name: String,
-	val accessToken: String?,
+	val accessToken: String? = null,
 )
