+++ app/src/main/java/org/jellyfin/androidtv/data/repository/LoginState.kt
@@ -1,0 +1,7 @@
+package org.jellyfin.androidtv.data.repository
+
+sealed class LoginState
+object AuthenticatingState : LoginState()
+object RequireSignInState : LoginState()
+object ServerUnavailableState : LoginState()
+object AuthenticatedState : LoginState()
