+++ app/src/main/java/org/jellyfin/androidtv/data/repository/ServerRepository.kt
@@ -29,17 +29,6 @@ interface ServerRepository {
 	fun addServer(address: String): Flow<ServerAdditionState>
 }
 
-sealed class ServerAdditionState
-object ConnectingState : ServerAdditionState()
-data class UnableToConnectState(val error: Exception) : ServerAdditionState()
-data class ConnectedState(val publicInfo: PublicSystemInfo) : ServerAdditionState()
-
-sealed class LoginState
-object AuthenticatingState : LoginState()
-object RequireSignInState : LoginState()
-object ServerUnavailableState : LoginState()
-object AuthenticatedState : LoginState()
-
 class ServerRepositoryImpl(
 	private val jellyfin: Jellyfin,
 	private val device: IDevice,
