+++ app/src/main/res/xml/authenticator.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<account-authenticator
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
     android:accountType="org.jellyfin.v1"
     android:icon="@drawable/ic_jellyfin"
-    android:smallIcon="@drawable/ic_jellyfin"
-    android:label="@string/app_name"/>
+    android:label="@string/app_name"
+    android:smallIcon="@drawable/ic_jellyfin" />
